Fix issues with checkout step enumeration
This required adding code inside of the checkout.setMethod method in
order to properly set the currentStep as 'login' since Magento never
actually does this and always sets the currentStep as 'billing'. Some
minor reorganization was done on the actual checkout.gotoSection
replacement to make sure that we're incrementing the step number (as
it's from a zero-indexed array), and also to make sure that we're
sending the *current* step as opposed to the step we're currently moving
to.