Overhaul the checkout page scripts
This adds a default initial checkout pageview and changes the general
setup of how the pageviews are created. Checkout pageviews are sent when
the step *starts* and also sends out some extra checkout_options for the
'login' and 'shipping_method' steps. This also removes some of the extra
parameters that were previously being sent along.