Remove setCacheLifetime from onestepcheckout block
Magento reads the null value as a string. Which sets a cache lifetime. This has the opposite effect as desired. By default, a core/template block doesn't have a cache lifetime anyway. Therefor, setting it is unneeded for the desired effect.

I confirmed this block was being cached by adding `<?php echo date('r'); ?>` to the end of the file, and seeing the timestamp never changed. Also adding/removing items from the cart would persist items on the checkout.